+++ src/index.ts
@@ -2,6 +2,7 @@ export * from "./types.js";
 export * from "./constants.js";
 export * from "./starter.js";
 export * from "./course.js";
+export * from "./course-v3.js";
 export * from "./validation.js";
 export * from "./engine.js";
 export * from "./assets.js";